Fix notification stacking
Replaces the previous implementation of notification
rendering. A bug would occur when a user finds two characters
fast, changing the state of a single notification component.
In an edge case where the Notification component would dismount
at ~ 4s, it would set the openNotification to false, causing
any immediate clicks to not render the notification.